bscWallet/api/requirements.txt
aiohttp==3.8.1
aiormq==6.2.3
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.4.1
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
base58==2.1.1
bcrypt==3.2.0
bitarray==1.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
cytoolz==0.11.2
ecdsa==0.17.0
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
fastapi==0.71.0
frozenlist==1.2.0
h11==0.12.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
mnemonic==0.20
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
pamqp==3.1.0
paramiko==2.9.2
parsimonious==0.8.1
pika==1.2.0
protobuf==3.19.3
pycparser==2.21
pycryptodome==3.12.0
pydantic==1.9.0
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dotenv==0.19.2
requests==2.27.1
rlp==2.0.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toolz==0.11.2
typing-extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.0
varint==1.0.2
web3==5.26.0
websockets==9.1
yarl==1.7.2
